Move tests on remote to test_..._remote.py
